requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
django-seed==0.3.1
Django==4.0.5
djangorestframework==3.13.1
Faker==13.13.0
pip==22.1.2
psycopg2-binary==2.9.3
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.1
setuptools==62.4.0
six==1.16.0
sqlparse==0.4.2
toposort==1.7
wheel==0.37.1
djangorestframework-simplejwt==5.2.0
django-filter~=21.1